Feature: responsive slider section [ref:no-ref]
