+++ src/index.scss
@@ -6,3 +6,20 @@ body {
   -webkit-font-smoothing: antialiased;
   -moz-osx-font-smoothing: grayscale;
 }
+
+@media (max-width: 980px) {
+  #root {
+    display: grid;
+    grid-template-areas:
+      "mainArea"
+      "headerArea";
+
+    header {
+      grid-area: headerArea;
+    }
+
+    main {
+      grid-area: mainArea;
+    }
+  }
+}
