+++ src/pages/MainPage/MainPage.module.scss
@@ -7,3 +7,38 @@
     height: 91vh;
   }
 }
+
+@media (max-width: 980px) {
+  .slider-section {
+    height: 100%;
+    background: url("../../assets/img/svg/header-bg.svg") top no-repeat;
+  }
+
+  .resize-wrap {
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    align-items: center;
+
+    & > div {
+      padding-top: 50px;
+      margin-bottom: 50px;
+
+      & > img {
+        width: 200px;
+      }
+    }
+  }
+}
+
+@media (max-width: 640px) {
+  .resize-wrap {
+    & > div {
+      padding-top: 25px;
+
+      & > img {
+        width: 92px;
+      }
+    }
+  }
+}
